Add all the well-known languages
cf. http://stevehardie.com/2009/10/list-of-common-languages/

"""
So, in collating this list of languages, I have decided to use the list of languages that Microsoft use, and double check it against Wikipedia’s list of most commonly used languages.
Secondly, I would like the database to store the information based on the ISO 639 code for languages (2 letters)
"""